refactor: rename vector index factory and add explicit backend validation
- Renamed `src/lib/vector-index/index.js` to `src/lib/vector-index/factory.js`.
- Updated `createVectorIndex` to explicitly validate the backend at the beginning of the function.
- Updated imports in `SemanticCache` and tests.
