+++ src/lib/vector-index/factory.js
@@ -15,6 +15,12 @@ export async function createVectorIndex(options = {}) {
   // Determine backend: environment variable > option > default
   const backend = process.env.VECTOR_INDEX_BACKEND || options.backend || 'annoy';
 
+  // Explicitly validate supported backends
+  const supportedBackends = ['annoy', 'hnsw'];
+  if (!supportedBackends.includes(backend)) {
+    throw new Error(`Unknown vector index backend: ${backend}. Use 'annoy' or 'hnsw'.`);
+  }
+
   if (backend === 'annoy') {
     // Dynamically import AnnoyVectorIndex
     const { default: AnnoyVectorIndex } = await import('./annoy-index.js');
@@ -34,8 +40,6 @@ export async function createVectorIndex(options = {}) {
       );
     }
   }
-
-  throw new Error(`Unknown vector index backend: ${backend}. Use 'annoy' or 'hnsw'.`);
 }
 
 export { VectorIndex };
