+++ tests/vector-index/factory.test.js
@@ -1,5 +1,5 @@
 import { describe, test, expect } from 'bun:test';
-import { createVectorIndex } from '../../src/lib/vector-index/index.js';
+import { createVectorIndex } from '../../src/lib/vector-index/factory.js';
 
 describe('VectorIndex Factory', () => {
   test('should create AnnoyVectorIndex by default', async () => {
@@ -23,8 +23,7 @@ describe('VectorIndex Factory', () => {
   });
 
   test('should throw error for unknown backend', async () => {
-    expect(async () => {
-      await createVectorIndex({ dim: 128, maxElements: 1000, backend: 'unknown' });
-    }).toThrow();
+    expect(createVectorIndex({ dim: 128, maxElements: 1000, backend: 'unknown' }))
+      .rejects.toThrow(/Unknown vector index backend/);
   });
 });
